sdk/requirements.txt
httpx>=0.24.0
langchain>=0.1.0
langchain-community>=0.0.1
pypdf>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
